pyproject.toml
[tool.poetry]
name = "vid2doc"
version = "0.1.0"
description = ""
authors = ["Rens Dimmendaal <rens@dimmendaal.com>"]
readme = "README.md"
packages = [{include = "vid2doc", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
pytube = "^15.0.0"
opencv-python = "^4.10.0.82"
numpy = "^1.26.4"
pillow = "^10.3.0"
scipy = "^1.13.1"
six = "^1.16.0"
imagehash = "^4.3.1"
tqdm = "^4.66.4"
openai = "^1.34.0"
typer = {extras = ["all"], version = "^0.12.3"}
youtube-transcript-api = "^0.6.2"
pydantic = "^2.7.4"
instructor = "^1.3.3"
python-dotenv = "^1.0.1"

[tool.poetry.scripts]
vid2doc = "vid2doc.cli:app"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
ruff = "^0.4.8"
jupyterlab = "^4.2.2"
ipywidgets = "^8.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
